+++ CHANGELOG.rst
@@ -7,6 +7,9 @@ Note that ``PHAB_ID=#`` and ``RB_ID=#`` correspond to associated messages in com
 Unreleased
 ----------
 
+20.4.1
+------
+
 New Features
 ~~~~~~~~~~~~
 * util-tunable: ConfigurationLinter accepts a relative path. ``PHAB_ID=D468528``